+++ eras/alonzo/impl/CHANGELOG.md
@@ -67,6 +67,7 @@
 
 ### `testlib`
 
+* Added `ShelleyEraImp` constraint to `emptyAlonzoImpNES`
 * Rename `emptyAlonzoImpNES` to `initAlonzoImpNES`
 * Remove `Test.Cardano.Ledger.Alonzo.CostModel` in favor of `Test.Cardano.Ledger.Plutus`
 * Move `costModelParamsCount` to `Cardano.Ledger.Plutus.CostModels`